refactor delete user functionality for the admin user

+++ api/controllers/user.controller.js
@@ -64,20 +64,18 @@ const updateUser = async (req, res, next) => {
 
 const deleteUser = async (req, res, next) => {
   if (!req.user.isAdmin && req.user.id !== req.params.userId) {
-    return next(
-      errorHandler(
-        403,
-        'You are not allowed to delete this user ' +
-          req.user.id +
-          ' ' +
-          req.params.userId
-      )
-    );
+    return next(errorHandler(403, 'You are not allowed to delete this user'));
   }
 
   try {
-    await User.findByIdAndDelete(req.params.userId);
-    res.status(200).json({ message: 'User has been deleted' });
+    const user = await User.findById(req.params.userId);
+
+    if (user.isAdmin && req.user.id !== req.params.userId) {
+      return next(errorHandler(403, 'You are not allowed to delete this user'));
+    } else {
+      await User.findByIdAndDelete(req.params.userId);
+      res.status(200).json({ message: 'User has been deleted' });
+    }
   } catch (error) {
     next(error);
   }
